LayoutTests/traversal/moz-bug559526-expected.txt

CONSOLE MESSAGE: Should have thrown an exception: InvalidStateError: Recursive filters are not allowed
CONSOLE MESSAGE: Should have thrown an exception: InvalidStateError: Recursive filters are not allowed
CONSOLE MESSAGE: Should have thrown an exception: InvalidStateError: Recursive filters are not allowed
CONSOLE MESSAGE: Should have thrown an exception: InvalidStateError: Recursive filters are not allowed
CONSOLE MESSAGE: Should have tests 4 filter calls: 4
Mozilla Bug 559526
